add click animation
